Distributed Systems/Problem 2/requirements.txt
psycopg2==2.7.4
PyYAML==3.12
